Simplify regex for N to N relations
This had me confused for a while. I was thinking the parenthesis were used to capture just the name. Turns out it's the entire "person:{name}" that's being used. I was thinking it was just "{name}" because of the parenthesis. It makes so much more sense now that I know it's not though.